pyproject.toml
[project]
name = "attention-lp"
version = "0.1.0"
description = "Interpretability-guided layer pruning for multilingual MT"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "bitsandbytes>=0.44.0",
    # Fine-tuning
    "peft>=0.13.0",
    "trl>=0.12.0",
    # Evaluation
    "unbabel-comet>=2.2.0",
    "sacrebleu>=2.4.0",
    # Inference
    "vllm>=0.6.0",
    # Data prep
    "sentence-transformers>=3.0.0",
    "fasttext-wheel>=0.9.2",
    "datasets>=3.0.0",
    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[project.scripts]
attention-lp = "src.cli:main"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
